+++ packages/liveblocks-react/src/__tests__/useThreads.test.tsx
@@ -24,7 +24,7 @@ const threads: ThreadData[] = [
     metadata: {},
     roomId: "room",
     type: "thread",
-    createdAt: "2021-10-06T01:45:56.558Z",
+    createdAt: new Date("2021-10-06T01:45:56.558Z"),
     comments: [],
   },
 ];